Crate hoover3_macro2

Implementation of the #[activity] and #[workflow] macros. If we implemented these in hoover_macro crate, a proc_macro crate, we couldn’t unit test their internals. So we have to implement them in a separate crate that’s not a proc_macro crate.

Re-exports

pub use syn;

Functions

activity

Macro extracts argument and return type from function definition, then generates the make_activity/_sync macro call.

extract_activity_arg_type 🔒

Get activity function argument type T from a function declared like this: #[activity] fn foo(x: T) -> anyhow::Result<V> { ... }

extract_activity_return_type 🔒

Get activity function return type V from a function declared like this: #[activity] fn foo(x: T) -> anyhow::Result<V> { ... }

extract_workflow_arg_type 🔒

Get workflow function argument type T from a function declared like this: #[workflow] async fn foo(ctx: WfContext, x: T) -> anyhow::Result<V> { ... }

extract_workflow_return_type 🔒

Get workflow function return type V from a function declared like this: #[workflow] async fn foo(ctx: WfContext, x: T) -> WorkflowResult<V> { ... }

workflow

Macro extracts argument and return type from function definition, then generates the make_workflow macro call.